requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=8.4.0
PyYAML>=5.3.1
scipy>=1.4.1  # Google Colab version
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
loguru

onnx>=1.9.0  # ONNX export
onnx-simplifier>=0.3.6  # ONNX simplifier
tensorflow==2.4.1  # TFLite export
tensorflow_datasets
tensorflowjs>=3.9.0  # TF.js export
openvino-dev  # OpenVINO export
openvino2tensorflow==1.31.3

onnxruntime-gpu
onnxruntime